requirements.txt
pytest==7.3.1
selene==2.0.0b17
allure-pytest==2.13.2
pip==23.1.2
